+++ solutions_builder/modules/terraform_base/terraform/stages/2-foundation/artifact_registry.tf
@@ -1,0 +1,6 @@
+resource "google_artifact_registry_repository" "default" {
+  location      = var.artifact_registry_region
+  repository_id = "default"
+  description   = "Default docker repository"
+  format        = "DOCKER"
+}